refactor: alter the structure to centrilize the handlers
