+++ internal/transport/rest/handler/helper.go
@@ -1,4 +1,4 @@
-package rest
+package handler
 
 import (
 	"encoding/json"
