+++ internal/transport/rest/router.go
@@ -1,8 +1,12 @@
 package rest
 
-import "net/http"
+import (
+	"net/http"
 
-func NewRouter(bookHandler BookHandler) http.Handler {
+	"github.com/CleysonPH/reading-tracker/internal/transport/rest/handler"
+)
+
+func NewRouter(bookHandler handler.BookHandler) http.Handler {
 	router := http.NewServeMux()
 	router.HandleFunc("/api/v1/books", bookHandler.GetBooks)
 	return router
